Add Pending{Name|DateOfBirth} properties to v3 teacher response
