+++ QualifiedTeachersApi/src/QualifiedTeachersApi/V3/Responses/GetTeacherResponse.cs
@@ -12,10 +12,12 @@ public record GetTeacherResponse
     [SwaggerSchema(Nullable = false)]
     public required string FirstName { get; init; }
     [SwaggerSchema(Nullable = false)]
-    public required string LastName { get; init; }
     public required string MiddleName { get; init; }
+    public required string LastName { get; init; }
     public required DateOnly DateOfBirth { get; init; }
     public required string NationalInsuranceNumber { get; init; }
+    public required bool PendingNameChange { get; init; }
+    public required bool PendingDateOfBirthChange { get; init; }
     public required GetTeacherResponseQts Qts { get; init; }
     public required GetTeacherResponseEyts Eyts { get; init; }
     public required GetTeacherResponseInduction Induction { get; init; }
